server/requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Auth & validation
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
pydantic>=2.0.0
bcrypt>=4.0.1

# Environment & HTTP
python-dotenv>=1.0.0
requests>=2.31.0

# AI / OpenRouter (copilot)
openai>=1.0.0
google-generativeai>=0.3.0

# RAG: FAISS vector store + embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Document processing (Knowledge Bridge)
PyPDF2>=3.0.0

# Fraud Detection: OCR with TrOCR
transformers>=4.30.0
torch>=2.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
